+++ comms/lighthouse/src/config/configService.ts
