+++ comms/lighthouse/src/config/default_layers.ts
