+++ comms/lighthouse/src/config/simpleStorage.ts
