+++ comms/lighthouse/src/misc/logging.ts
