+++ comms/lighthouse/src/peers/archipelagoService.ts
