+++ comms/lighthouse/src/peers/idService.ts
