+++ comms/lighthouse/src/server.ts
@@ -8,17 +8,17 @@ import express from 'express'
 import morgan from 'morgan'
 import * as path from 'path'
 import { IRealm } from 'peerjs-server'
-import { ArchipelagoService } from './archipelagoService'
-import { ConfigService, LighthouseConfig } from './configService'
-import { DEFAULT_LAYERS } from './default_layers'
-import { IdService } from './idService'
-import { initPeerJsServer } from './initPeerJsServer'
+import { ConfigService, LighthouseConfig } from './config/configService'
+import { DEFAULT_LAYERS } from './config/default_layers'
+import { lighthouseConfigStorage } from './config/simpleStorage'
 import { LayersService } from './layersService'
-import { patchLog } from './logging'
-import { pickName } from './naming'
-import { PeersService } from './peersService'
+import { patchLog } from './misc/logging'
+import { pickName } from './misc/naming'
+import { ArchipelagoService } from './peers/archipelagoService'
+import { IdService } from './peers/idService'
+import { initPeerJsServer } from './peers/initPeerJsServer'
+import { PeersService } from './peers/peersService'
 import { configureRoutes } from './routes'
-import { lighthouseConfigStorage } from './simpleStorage'
 
 const LIGHTHOUSE_VERSION = '0.2'
 const DEFAULT_ETH_NETWORK = 'ropsten'
