--- comms/lighthouse/src/storageKeys.ts
@@ -1,3 +1,0 @@
-export const StorageKeys = {
-  PEER_TOKEN: 'peerToken'
-}
