+++ comms/lighthouse/test/naming.spec.ts
@@ -5,7 +5,7 @@ process.env.LIGHTHOUSE_STORAGE_LOCATION = '.'
 import { DAOClient } from 'decentraland-katalyst-commons/DAOClient'
 import { ServerMetadata } from 'decentraland-katalyst-commons/ServerMetadata'
 import { defaultNames, pickName } from '../src/naming'
-import { lighthouseStorage } from '../src/simpleStorage'
+import { lighthouseStorage } from '../src/config/simpleStorage'
 
 declare let global: any
